Added support for BSP patch files
When loading maps/onasi_is_easy.bsp, server will look for maps/onasi_is_easy.bsp.patched and use that one if it exists.

The map checksum will still be calculated with the non-patched version. This allows server owners to have a base folder with clean pk3s for basejka auto dl purposes (and generally as a better habit) while being able to put up hotfix for maps.

Patch files can be loaded as loose (unpacked) files in the filesystem or if you have multiple, you can put them all in a single pk3 (not in the original map pk3 though as it breaks the checksum).